changed some long types to size_t

+++ include/program/jobs/Compute_Morse_Decomposition.h
@@ -27,10 +27,10 @@
 /// path possible."
 template < class Toplex , class Conley_Morse_Graph , class Combinatorial_Map >
 void Compute_Morse_Decomposition ( Conley_Morse_Graph * conley_morse_graph ,
-  std::map < typename Conley_Morse_Graph::Vertex , long > * exit_path_bounds ,
-  std::map < typename Conley_Morse_Graph::Vertex , long > * entrance_path_bounds ,
+  std::map < typename Conley_Morse_Graph::Vertex , size_t > * exit_path_bounds ,
+  std::map < typename Conley_Morse_Graph::Vertex , size_t > * entrance_path_bounds ,
   std::map < typename Conley_Morse_Graph::Edge , typename Toplex::Toplex_Subset > * connecting_orbits ,
-  std::map < typename Conley_Morse_Graph::Edge , long > * path_bounds ,
+  std::map < typename Conley_Morse_Graph::Edge , size_t > * path_bounds ,
   long * through_path_bound ,
   const typename Toplex::Toplex_Subset & domain ,
   const Combinatorial_Map & combinatorial_map );
